refactor: item 내용 삭제에 관한 컴포넌트 분리
